backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
motor==3.6.0
pymongo==4.10.1
python-dotenv==1.0.1
pydantic==2.9.2
httpx==0.27.2
